Support complex toolchains
Instead of specifying the name of the toolchain and relying on arm-eabi-
use the entire path to the custom toolchain.

Add the following line to BoardConfig.mk to use the custom toolchain for kernel
For linaro:
TARGET_KERNEL_CUSTOM_TOOLCHAIN := linaro/bin/arm-linux-gnueabihf-

For GCC 4.4.3:
TARGET_KERNEL_CUSTOM_TOOLCHAIN := arm-eabi-4.4.3/bin/arm-eabi-

Add the following to slim.dependencies to get the toolchain project.

{
    "repository": "mnm9994u/prebuilt",
    "target_path": "prebuilt",
    "branch": "master"
},

Or add the following to your local_manifest

<project name="mnm9994u/prebuilt" path="prebuilt" remote="github" revision="master" />

Change-Id: I45ed019278a1f310d6abc4c7fbfa14ddc6d184b7