Fully expand eng/debug/tests module install files
When including eng, debug and tests modules that are tagged
using LOCAL_MODULE_TAGS, fully expand the install file list
from just the tagged module to all the module's INSTALLED
list.  This allows auto-generated files such as dex preopt
odex files tagged under the module's INSTALLED list to be
included.  This also makes the behavior of specifying a
module as debug using LOCAL_MODULE_TAGS to be the same as
specifying it as PRODUCT_PACKAGES_DEBUG when building
for userdebug or eng variant.

Change-Id: If9a24b62d8bee9e378f3e5f58e9490e963238302