core: Makefile: Build install-recovery into the system image
We now generate and install install-recovery.sh and recovery-
from-boot.p as a part of the normal make. This is needed in
order for incremental ota updates to work. Without this change
the targetfiles.zip which is used as the source for the incremental
package contains the two files from above but the system image
generated by the same build does not.This causes the update to fail
when we try to apply it

Change-Id: Ibe79eb8f19e89a775f72a330bea1ed5552109f4f

Conflicts:
	core/Makefile