Update patch set 1
Patch Set 1:

Correct it is for debugging. and since its not enabled in user builds thats why I turned it off here. I have not noticed any issues with it being off. it is meant to catch bugs in kernel but we really dont have any bugs known in the mainline kernels we use that this would even be remotely helpful with

Patch-set: 1